// src/types/subscription.ts
export enum SubscriptionFeatureEnum {
  SUB_ACCOUNTS_3 = '3 Sub accounts',
  TEAM_MEMBERS_2 = '2 Team members',
  UNLIMITED_PIPELINES = 'Unlimited pipelines',
  REBILLING = 'Rebilling',
  SUPPORT_TEAM_24_7 = '24/7 Support team',
  UNLIMITED_SUB_ACCOUNTS = 'Unlimited Sub accounts',
  UNLIMITED_TEAM_MEMBERS = 'Unlimited Team members',
}

export enum SubscriptionTitleEnum {
  STARTER = 'starter',
  UNLIMITED_SAAS = 'unlimited-saas',
  BASIC = 'basic',
}

export enum SubscriptionDurationEnum {
  MONTH = 'month',
  YEAR = 'year',
}